gate-alpha-agent/requirements.txt
# Gate Alpha Agent - Requirements

# Core dependencies
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
structlog>=23.2.0
ollama>=0.4.0

# Gate.io API SDK
gate-api>=4.36.0

# Async support (included in Python 3.11+)
# asyncio

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Type checking (optional, for development)
# mypy>=1.7.0
